Fixed infinite recursion

See merge request pap-z21/cPAP2023Z-Z21!15

+++ src/main/java/pw/pap/service/ProjectService.java
@@ -1,5 +1,7 @@
 package pw.pap.service;
 
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,6 +13,7 @@ import pw.pap.repository.UserRepository;
 import pw.pap.repository.ProjectRepository;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -27,14 +30,22 @@ public class ProjectService {
         this.taskRepository = taskRepository;
     }
 
-    public Project createProject(String name, String description, LocalDateTime projectDeadline, User owner, List<User> members) {
+    public Project createProject(String name, User owner) {
         LocalDateTime currentDate = LocalDateTime.now();
-        Project project = new Project(name, description, currentDate, projectDeadline, owner, members);
+        Project project = new Project(name, owner, currentDate);
         return projectRepository.save(project);
     }
 
-    public Project addProject(Project project) {
-        return projectRepository.save(project);
+    public List<Task> getProjectTasks(Long projectId){
+        List<Task> projectTasks = new ArrayList<>();
+        Iterable<Task> tasks = taskRepository.findAll();
+
+        for (Task task : tasks){
+            if(task.getProject().getId().equals(projectId)){
+                projectTasks.add(task);
+            }
+        }
+        return projectTasks;
     }
 
     public Optional<Project> getProjectById(Long projectId) {
@@ -46,29 +57,49 @@ public class ProjectService {
     }
 
     public Project updateProject(Long projectId, Project updatedProject) {
+        LocalDateTime currentDate = LocalDateTime.now();
         Project existingProject = projectRepository.findById(projectId)
             .orElseThrow(() -> new IllegalArgumentException("Project not found"));
 
-        projectRepository.deleteById(projectId);
-        updatedProject.setId(projectId);
-        return projectRepository.save(updatedProject);
+        String newName = updatedProject.getName();
+        if (newName != null && !newName.equals(existingProject.getName())){
+            if (updatedProject.getName().isBlank()) {
+                throw new IllegalArgumentException("Project name cannot be empty");
+            }
+            existingProject.setName(newName);
+        }
+
+        String newDescription = updatedProject.getDescription();
+        if (newDescription != null) {
+            existingProject.setDescription(newDescription);
+        }
+
+        LocalDateTime newDeadline = updatedProject.getProjectDeadline();
+        if (newDeadline != null) {
+            if (!newDeadline.isAfter(currentDate)) {
+                throw new IllegalArgumentException("New deadline must be after current time");
+            }
+            existingProject.setProjectDeadline(newDeadline);
+        }
+
+        if (updatedProject.getOwner() != null) {
+            User newOwner = userRepository.findById(updatedProject.getOwner().getId())
+                .orElseThrow(() -> new EntityNotFoundException("New owner not in database"));
+            existingProject.setOwner(newOwner);
+        }
+
+        return projectRepository.save(existingProject);
     }
 
     public void deleteProject(Long projectId) {
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(() -> new IllegalArgumentException("Project not found"));
 
-        User owner = project.getOwner();
-        owner.getOwnedProjects().remove(project);
-
-        List<User> members = project.getMembers();
-        for (User member : members) {
-            member.getMemberOfProjects().remove(project);
-        }
-
-        List<Task> tasks = project.getTasks();
+        Iterable<Task> tasks = taskRepository.findAll();
         for (Task task : tasks) {
-            taskRepository.deleteById(task.getId());
+            if(task.getProject().getId().equals(projectId)){
+                taskRepository.deleteById(task.getId());
+            }
         }
 
         projectRepository.deleteById(projectId);
@@ -82,8 +113,6 @@ public class ProjectService {
         User user = userRepository.findById(userId)
             .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        user.getMemberOfProjects().add(project);
-        userRepository.save(user);
         project.getMembers().add(user);
         projectRepository.save(project);
     }
@@ -96,12 +125,12 @@ public class ProjectService {
         User user = userRepository.findById(userId)
             .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        List<Task> assignedTasks = user.getAssignedTasks();
-        for (Task task : assignedTasks) {
-            if (task.getProject().equals(project)) {
-                task.getAssignees().remove(user);
-            }
-        }
+        Iterable<Task> tasks = taskRepository.findAll();
+         for (Task task : tasks) {
+             if (task.getProject().getId().equals(projectId)) {
+                 task.getAssignees().remove(user);
+             }
+         }
 
         project.getMembers().remove(user);
         projectRepository.save(project);
@@ -115,15 +144,6 @@ public class ProjectService {
         Task task = taskRepository.findById(taskId)
                 .orElseThrow(() -> new IllegalArgumentException("Task not found"));
 
-        User creator = task.getCreator();
-        creator.getCreatedTasks().remove(task);
-
-        List<User> assignees = task.getAssignees();
-        for (User assignee : assignees) {
-            assignee.getAssignedTasks().remove(task);
-        }
-
-        project.getTasks().remove(task);
         taskRepository.deleteById(taskId);
         projectRepository.save(project);
     }
